+++ src/fields/Matrix.php
@@ -44,6 +44,7 @@ use craft\validators\ArrayValidator;
 use craft\web\assets\matrix\MatrixAsset;
 use craft\web\assets\matrixsettings\MatrixSettingsAsset;
 use GraphQL\Type\Definition\Type;
+use Illuminate\Support\Collection;
 use yii\base\InvalidArgumentException;
 use yii\base\InvalidConfigException;
 use yii\db\Expression;
@@ -548,7 +549,7 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
      */
     public function serializeValue(mixed $value, ?ElementInterface $element = null): mixed
     {
-        /** @var MatrixBlockQuery $value */
+        /** @var MatrixBlockQuery|Collection $value */
         $serialized = [];
         $new = 0;
 
@@ -756,7 +757,7 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
      */
     public function isValueEmpty(mixed $value, ElementInterface $element): bool
     {
-        /** @var MatrixBlockQuery $value */
+        /** @var MatrixBlockQuery|Collection $value */
         return $value->count() === 0;
     }
 
@@ -767,7 +768,7 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
      */
     public function validateBlocks(ElementInterface $element): void
     {
-        /** @var MatrixBlockQuery $value */
+        /** @var MatrixBlockQuery|Collection $value */
         $value = $element->getFieldValue($this->handle);
         $blocks = $value->all();
         $allBlocksValidate = true;
@@ -822,7 +823,7 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
      */
     protected function searchKeywords(mixed $value, ElementInterface $element): string
     {
-        /** @var MatrixBlockQuery $value */
+        /** @var MatrixBlockQuery|Collection $value */
         $keywords = [];
 
         foreach ($value->all() as $block) {
@@ -843,7 +844,7 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
      */
     public function getStaticHtml(mixed $value, ElementInterface $element): string
     {
-        /** @var MatrixBlockQuery $value */
+        /** @var MatrixBlockQuery|Collection $value */
         $value = $value->all();
 
         /** @var MatrixBlock[] $value */
@@ -1059,10 +1060,12 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
 
         // Repopulate the Matrix block query if this is a new element
         if ($resetValue || $isNew) {
-            /** @var MatrixBlockQuery $query */
-            $query = $element->getFieldValue($this->handle);
-            $this->_populateQuery($query, $element);
-            $query->clearCachedResult();
+            /** @var MatrixBlockQuery|Collection $value */
+            $value = $element->getFieldValue($this->handle);
+            if ($value instanceof MatrixBlockQuery) {
+                $this->_populateQuery($value, $element);
+            }
+            $value->clearCachedResult();
         }
 
         parent::afterElementPropagate($element, $isNew);